refactor(chat): extract runner contracts
